refactor(root): address Greptile P2s — dedupe letterbox into fork, fix even() rounding, warn on dropped pad
Move computeLetterbox + types into @shepherdjerred/discord-video-stream (both bots
already depend on it) and delete the byte-identical per-bot copies. Fix even() to
round to the nearest even integer (was always rounding up) and correct its doc.
Emit a debug-level warning when prepareStream's pad is set while a hardware decode
pipeline is active (scale_vaapi ignores pad) instead of silently dropping it.
