+++ packages/discord-plays-mario-kart/packages/backend/eslint.config.ts
@@ -10,7 +10,6 @@ const config = [
         "src/game/command/command-input.test.ts",
         "src/emulator/buttons.test.ts",
         "src/emulator/png.test.ts",
-        "src/stream/letterbox.test.ts",
       ],
     },
   }),
