+++ packages/discord-video-stream/src/media/index.ts
@@ -1,4 +1,5 @@
 export * from "./LibavDemuxer.js";
+export * from "./letterbox.js";
 export * from "./newApi.js";
 export * as NewApi from "./newApi.js";
 export * from "./encoders/index.js";
